ielts_crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
lxml>=5.1.0
html5lib>=1.1
playwright>=1.40.0
